backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
SQLAlchemy>=2.0.30
alembic>=1.13.2
pydantic-settings>=2.3.4
python-dotenv>=1.0.1
trafilatura>=1.12.2
readability-lxml>=0.8.1
beautifulsoup4>=4.12.3
pytest>=8.2.0
